Extract metadata set/get functions to yamlmeta.go
- make more explicit this is the extension of the yamlmeta "meta"
  feature.
